requirements.txt
mysql-connector-python == 9.1.0
streamlit == 1.40.2

